Move inset function to inset.py
