+++ pygmt/src/__init__.py
@@ -1,5 +1,7 @@
 """
 Source code for PyGMT modules.
 """
+from pygmt.src.inset import inset
+
 # pylint: disable=import-outside-toplevel
 from pygmt.src.meca import meca
